Switch to using mdnsjs as the first available lib
Testing the waters to see if switching to mdns-js as primary will
be as detrimental as we originally thought. The on going thought is
that machines with bonjour/avahi already installed will cause issues
with the JS version, however I tested on my system and it seems to
work without error.